# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Project Dashboard Bridge",
    "description": "Read and write to the Project Dashboard. Use get_projects to list projects, get_project to see tasks, create_task to add work, log_execution to record results, update_task to change status. You must log_execution before marking a task done or failed.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://project-dashboard-aj.vercel.app"
    }
  ],
  "paths": {
    "/api/bridge/projects": {
      "get": {
        "operationId": "get_projects",
        "summary": "List all active projects with task counts",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" },
                    "projects": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "integer" },
                          "name": { "type": "string" },
                          "description": { "type": "string" },
                          "status": { "type": "string" },
                          "project_url": { "type": "string" },
                          "project_type": { "type": "string" },
                          "completed_tasks": { "type": "integer" },
                          "total_tasks": { "type": "integer" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/bridge/projects/{id}": {
      "get": {
        "operationId": "get_project",
        "summary": "Get one project with all its tasks",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "Project ID"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" },
                    "project": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "integer" },
                        "name": { "type": "string" },
                        "description": { "type": "string" },
                        "status": { "type": "string" },
                        "project_url": { "type": "string" },
                        "project_type": { "type": "string" },
                        "tasks": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": { "type": "integer" },
                              "title": { "type": "string" },
                              "description": { "type": "string" },
                              "status": { "type": "string" },
                              "execution_prompt": { "type": "string" },
                              "assigned_agent": { "type": "string" },
                              "priority": { "type": "string" }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/bridge/tasks": {
      "post": {
        "operationId": "create_task",
        "summary": "Create a new task in a project",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["project_id", "title"],
                "properties": {
                  "project_id": { "type": "integer", "description": "Project ID to add the task to" },
                  "title": { "type": "string", "description": "Task title" },
                  "description": { "type": "string", "description": "Task description" },
                  "execution_prompt": { "type": "string", "description": "Instructions for the agent executing this task" },
                  "assigned_agent": { "type": "string", "description": "Agent name: planner, developer, tester, or chatgpt" },
                  "priority": { "type": "string", "description": "low, medium, or high" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" },
                    "task": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "integer" },
                        "project_id": { "type": "integer" },
                        "title": { "type": "string" },
                        "status": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/bridge/tasks/{id}": {
      "patch": {
        "operationId": "update_task",
        "summary": "Update a task. To mark done or failed you must call log_execution first.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "Task ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": { "type": "string", "description": "New status: todo, in-progress, done, failed, or pending" },
                  "title": { "type": "string", "description": "New title" },
                  "description": { "type": "string", "description": "New description" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Must log execution before marking done or failed"
          }
        }
      }
    },
    "/api/bridge/execution-logs": {
      "post": {
        "operationId": "log_execution",
        "summary": "Log an execution result for a task. Required before marking a task done or failed.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["task_id", "agent", "result"],
                "properties": {
                  "task_id": { "type": "integer", "description": "Task ID this log belongs to" },
                  "agent": { "type": "string", "description": "Who executed: chatgpt, kiro, developer, tester" },
                  "result": { "type": "string", "description": "What was done or the outcome" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Execution logged",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" },
                    "log_id": { "type": "integer" },
                    "task_id": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BridgeKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Bridge-Key"
      }
    }
  },
  "security": [
    { "BridgeKey": [] }
  ]
}